refactor(components)
- Small shuffling to get top-levle components and modules dirs
